+++ concept/tests/fluid_drift_rigid_nohubble/generate_IC.py
@@ -13,15 +13,15 @@ gridsize = 64
 Vcell = (boxsize/gridsize)**3
 speed = boxsize/(10*units.Gyr)
 component = Component('test fluid', 'matter', gridsize=gridsize)
-ϱ = empty([gridsize]*3)
+ϱ = empty([gridsize]*3, dtype=float)
 for i in range(gridsize):
-    ϱ[i, :, :] = 2 + np.sin(2*π*i/gridsize)  # Unitless
-ϱ /= sum(ϱ)                                  # Normalize
-ϱ *= ρ_mbar*gridsize**3                      # Apply units
-component.populate(ϱ,                             'ϱ'   )
-component.populate(ϱ*speed,                       'J', 0)
-component.populate(zeros([gridsize]*3),           'J', 1)
-component.populate(ϱ*speed*random_uniform(-1, 1), 'J', 2)
+    ϱ[i, :, :] = 2 + sin(2*π*i/gridsize)  # Unitless
+ϱ /= sum(ϱ)                               # Normalize
+ϱ *= ρ_mbar*gridsize**3                   # Apply units
+component.populate(ϱ,                                'ϱ'   )
+component.populate(ϱ*speed,                          'J', 0)
+component.populate(zeros([gridsize]*3, dtype=float), 'J', 1)
+component.populate(ϱ*speed*random_uniform(-1, 1),    'J', 2)
 
 # Save snapshot
 save(component, initial_conditions)
